\name{tnam-package}
\docType{package}
\alias{tnam-package}
\title{Temporal Network Autocorrelation Models (TNAM)}
\description{Temporal Network Autocorrelation Models (TNAM).}
\details{
The \pkg{tnam} package implements temporal and cross-sectional network autocorrelation models (TNAM).
}
\author{
Philip Leifeld (\url{http://www.philipleifeld.com})

Skyler J. Cranmer (\url{http://www.skylercranmer.net})
}
